Extract uniq name to common base
