+++ common/base/Cargo.toml
@@ -10,14 +10,15 @@ edition = "2021"
 # Workspace dependencies
 common-tracing = { path = "../tracing" }
 common-exception = { path = "../exception" }
-futures = "0.3"
-async-trait = "0.1"
-ctrlc = { version = "3.1.9", features = ["termination"] }
-pprof = { version = "0.5", features = ["flamegraph", "protobuf"] }
 
 # Github dependencies
 
 # Crates.io dependencies
+futures = "0.3"
+async-trait = "0.1"
+uuid = { version = "0.8", features = ["serde", "v4"] }
+ctrlc = { version = "3.1.9", features = ["termination"] }
+pprof = { version = "0.5", features = ["flamegraph", "protobuf"] }
 tokio = { version = "1.12.0", features = ["macros", "rt", "rt-multi-thread", "sync", "fs"] }
 
 [dev-dependencies]
