+++ dfs/src/tests/service.rs
@@ -17,6 +17,7 @@ use std::sync::Arc;
 use anyhow::Result;
 use common_base::tokio;
 use common_base::tokio::sync::oneshot;
+use common_base::GlobalSequence;
 use common_sled_store::get_sled_db;
 use common_tracing::tracing;
 use kvsrv::meta_service::GetReq;
@@ -54,7 +55,7 @@ pub async fn start_store_server_with_context(tc: &mut StoreTestContext) -> Resul
 }
 
 pub fn next_port() -> u32 {
-    19000u32 + (common_base::uniq_usize() as u32)
+    19000u32 + (GlobalSequence::next() as u32)
 }
 
 pub struct StoreTestContext {
